Simplify implementation and API
